rewriting tests small bug fixes
