+++ app/assets/javascripts/helpers/FromPair.js
