+++ app/assets/javascripts/student_profile/page_container.jsx
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import MixpanelUtils from '../helpers/mixpanel_utils.jsx';
 import PropTypes from '../helpers/prop_types.jsx';
-import {fromPair} from '../helpers/from_pair.jsx';
+import {fromPair} from '../helpers/FromPair.js';
 import {merge} from '../helpers/react_helpers.jsx';
 
 (function() {
