+++ spec/javascripts/student_profile/services_list_spec.jsx
@@ -1,6 +1,6 @@
 import {studentProfile} from './fixtures.jsx';
 import SpecSugar from '../support/spec_sugar.jsx';
-import {fromPair} from '../../../app/assets/javascripts/helpers/from_pair.jsx';
+import {fromPair} from '../../../app/assets/javascripts/helpers/FromPair.js';
 import {merge} from '../../../app/assets/javascripts/helpers/react_helpers.jsx';
 
 describe('ServicesList', function() {
